requirements.txt
marshmallow==3.20.2
openfabric_pysdk==0.2.7
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2.post1